Fix some rubyisms in the Character class
There are several rubyisms that were being violated in the Character
class. For example, the use of the keyword `return` is not necessary
since the last line inside a method will be returned. The convention of
prefixing method names with, e.g., `get_` and `set_` is not necessary in
Ruby source and you should prefer use of the `attr_*` methods instead.

Adds some spec coverage for the refactored Character object.